Commented out old API calls
Since all API calls are out of date, this version removes them and loads
the calls necessary to compile the plugin only.

This allows the new API calls to emulate the function of what the old
API did without the old API getting in the way. Old API calls are kept
for reference.